Moved key functions
Moved key functions into key_conversions.py file to separate functionality
