+++ main.py
@@ -1,51 +1,17 @@
 import spotipy
 from spotipy.oauth2 import SpotifyClientCredentials
 import pandas as pd
+import key_conversions
 import keys # Remove import
 
 # Define constants - set own credentials upon download
-SPOTIPY_CLIENT_ID = keys.SPOTIPY_CLIENT_ID
-SPOTIPY_CLIENT_SECRET = keys.SPOTIPY_CLIENT_SECRET
-PLAYLIST_ID = 'spotify:playlist:7wsEUldS5ZU2UDAMZealYG'
+SPOTIPY_CLIENT_ID = keys.SPOTIPY_CLIENT_ID # Set own Spotify Client ID here
+SPOTIPY_CLIENT_SECRET = keys.SPOTIPY_CLIENT_SECRET # Set own Spotify Client Secret
+PLAYLIST_ID = 'spotify:playlist:7wsEUldS5ZU2UDAMZealYG' # Set own Spotify Playlist ID
 
+# Constants used when querying API
 AUDIO_FEATURES_QUERY_SIZE = 100
 TRACKS__QUERY_SIZE = 50
-
-def keyIntToNote(item):
-    if item["key"] == 0:
-        item["key"] = "C"
-    elif item["key"] == 1:
-        item["key"] = "C#/D♭"
-    elif item["key"] == 2:
-        item["key"] = "D"
-    elif item["key"] == 3:
-        item["key"] = "D#/E♭"
-    elif item["key"] == 4:
-        item["key"] = "E"
-    elif item["key"] == 5:
-        item["key"] = "F"
-    elif item["key"] == 6:
-        item["key"] = "F#/G♭"
-    elif item["key"] == 7:
-        item["key"] = "G"
-    elif item["key"] == 8:
-        item["key"] = "G#/A♭"
-    elif item["key"] == 9:
-        item["key"] = "A"
-    elif item["key"] == 10:
-        item["key"] = "A#/B♭"
-    elif item["key"] == 11:
-        item["key"] = "B"
-
-    return item
-
-def modeIntToNote(item):
-    if item["mode"] == 0:
-        item["mode"] = "Minor"
-    elif item["mode"] == 1:
-        item["mode"] = "Major"
-
-    return item
 
 # Create auth manager and Spotify class for handing API calls
 auth_manager = SpotifyClientCredentials(client_id=SPOTIPY_CLIENT_ID, client_secret=SPOTIPY_CLIENT_SECRET)
@@ -104,8 +70,8 @@ for track_features in audio_features:
     track_features = eval(str(track_features))
     item = {"key" : track_features['key'], "mode" : track_features['mode'], "tempo" : track_features['tempo']}
 
-    item = keyIntToNote(item)
-    item = modeIntToNote(item)
+    item = key_conversions.keyIntToNote(item)
+    item = key_conversions.modeIntToNote(item)
     
     audio.append(item)
